backend/requirements.txt
# InsureGuard AI - Python Dependencies
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Validation
pydantic[email-validator]==2.7.0
email-validator==2.1.1

# ML/AI
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
xgboost==2.0.3
imbalanced-learn==0.12.3

# Document Processing
Pillow==10.3.0

# Environment
python-dotenv==1.0.1

# PDF Generation
fpdf2==2.7.9

# Utilities
httpx==0.27.0
